[#214] Fix Travis config
Problem: we do not explicitly specify which Ubuntu version should
be used by Travis and previously it was using Trusty, but now it seems
to be using Xenial, at least in some builds.
Some info can be found in this blog:
https://blog.travis-ci.com/2019-04-15-xenial-default-build-environment
We install z3 from ppa:hvr/z3, but it doesn't have a distribution for
Xenial (http://ppa.launchpad.net/hvr/z3/ubuntu/dists/).
So our Travis config is broken when launched on Xenial, but works on
Trusty.
In general not specifying Ubuntu version is fragile and leads to such
problems.
Solution:
1. Specify `dist: xenial` to avoid such breakages in the future.
2. Do not add ppa:hvr/z3, but instead install z3 from main Ubuntu
repo, it seems to be there since Xenial.
3. Replace `-qq` with `-q` and remove `--force-yes`, using those
is apparently not recommended by manual.
4. Install z3 only when liquidhaskell is going to be launched.
Linux check was not necessary, because apparently we always run it on
Linux.
5. Replace backticks with quotes in some messages, because otherwise
`liquidhaskell` in backticks means that `liquidhaskell` should be
launched, but it doesn't exist at those places.