[#226] Use GitHub Actions as the only CI
Problem: we are currently using two CI systems and they are both
external from GitHub. But now GitHub Actions is available.

Solution: take the config described in
https://kodimensional.dev/github-actions.
Remove travis and appveyor configuration.
Note that now we are building only one version with `stack` on CI.
Building with all supported GHCs is done only with `cabal`.
As a consequence, we don't need `stack-8*.yaml` files anymore.