[#234] Add concurrent primitives working with State
Problem: existing `modifyTVar'` is not pretty convenient. You need to
e.g. remember the order of arguments, and if several transformations
are required, composing them neatly may be difficult.

Solution: provide methods working with `State`, not a lambda. This way
one can write down modifications in do-block which is convenient.