[#239] Add a note on checking our text `RULE`
Problem: I've added a rewrite rule that replaces `toString . toText`
with `id`, however, it has to be checked each time `text` package is
updated, and this turns out to happen relatively often.
I had a plan on how to check that quickly, but didn't write it down,
which was wrong.

Solution: write down an easy way to check the comment validity.

Method that affect our rule are scattered across many modules, so it's
not very viable to mention particular module names. Instead, I refer the
reader to diff in `text` package, most likely it touches unrelated
things and then further check is not necessary.