Add special grouping functions
Problem: it is an often need to have a function of
`[(a, b)] -> [(a, [b])]` signature that would group elements by the
first element of the pair, and return the element we grouped by and the
associated second elements of the pairs.

Solution: add this function, and a more generic one that does not
necessarily work with pairs.

Also, we in-parallel work on making `group` function return `NonEmpty`,
not lists, so I implement it via `NonEmpty` here too.

Special thanks to @treeowl for providing elaborate implementations and
advanced variations for special cases.