Do not build universum-doctest with GHC >= 9.0.0
Problem: universum-doctest passes with GHC < 9, but starts to
fail with >= 9. The errors are weird, it complains that a lot of
stuff is not in scope. Example:
> Not in scope: type constructor or class ‘Text’

Solution: there is the following issue
https://github.com/sol/doctest/issues/327. Apparently, something is
wrong in `doctest` itself. The maintainer intends to investigate it,
so let's wait for that. While waiting, let's disable
`universum-doctest` whenever GHC is >= 9.0.0.