[#270] Use `tasty-discover`
Problem: In order to be compliant with the stackage nightly, the test
suite needs to buildable with tasty-hedgehog 1.2.0.0. At the moment, we
have a constraint on `tasty-hedgehog < 1.2.0.0` because that version
deprecates the `testProperty` function (and encourages using
`testPropertyNamed` instead), which we're still using.

Solution: we don't actually need to use either `testProperty` or
`testPropertyNamed`, we can just use `tasty-discover` to collect all
tests.