[Chore] Use a single PR template
Problem: GitHub's support for **multiple** PR templates is somewhat
lacking.

1. When you create an issue, you're presented with a screen to pick
   which template you want to use. However, when you create a PR, you're
   not. GH automatically uses a default template if there is one, or no
   template if there is no default.
2. If you want to use one of your PR templates, you **have** to manually
   add the `?template=file.md` query parameter to the URI (see docs:
   https://docs.github.com/en/pull-requests/collaborating-with-pull-requests/proposing-changes-to-your-work-with-pull-requests/using-query-parameters-to-create-a-pull-request),

One idea could be to have a default PR template for regular PRs, and
then a second custom one for release PRs. And in `CONTRIBUTING.md` we
could add a link to a custom URI to create a release PR, using the query
parameter as described above. However, this doesn't work either, because
in order to create the URI you have to already know which branch you
want to merge, e.g.:

```
https://github.com/serokell/universum/compare/master...<target-branch>?template=release.md
```

Since this is not known a priori, adding a link for release PRs to
`CONTRIBUTING.md` is not possible either.

Solution: Let's go back to using a single PR template. At the bottom,
we'll add an additional checklist for release-related PRs.