[#285] Update the comment in tests
Problem: after the large rewrite of `text` package (where it started
keeping UTF-8 instead of UTF-16) some things changed.

E.g. the code point mentioned in the comment is different now, and the
link to the `text`'s source code is not more actual.

Solution: strip the details that has to be maintained.

It all seems pretty clear without them.