[Chore] Prepare a release
Problem: `universum` has been excluded from the recent LTS, at least
because we do not support the most recent version of `text` package.
This issue with `text` has been fixed, and we need a new version with
the fix.

Solution: bump the version from `1.8.1.2` to `1.8.2`, since among
changes in definitions we only started exporting `~`.