[Chore] Mention `Unreleased` section in PR template
Problem: it's not really fixed anywhere, whether should we have an empty
`Unreleased` section in our changelog, or better go without it.

Solution: we agreed with @gromak that it's better to leave no empty
section (so that it does not appear on Hackage).

So as a clarification, on our PR template I change the point about
changelog (for regular PRs, not as part of release routine) to mention
the creation of `Unreleased` section if it is absent.