[Chore] Drop ghc-8.8 support
Problem: supporting ghc-8.8 forces us to have some conditionals
in the .cabal file. Not a big deal, but still some overhead.
Nowadays this compiler version is rather old, and since we claim
to always support 3 versions (as a minimum) it's safe to drop
the oldest supported one.

Solution: remove it from CI config and tested-with, remove
related conditionals.