[Chore] Relax the constraint on text
Problem: text-2.1.2 is available, but our constraints don't allow it.

Solution: we started have such a strict constraint on `text` when
we added a rewrite rule that relied on the implementation details
of `Data.Text.pack`. Because even a minor version bump could change
this detail. However, at some point the implementation had changed,
so this rewrite rule became unnecessary.
Now we have this rule only for old GHC versions, so this constraint
exists for a rule that is not even enabled, it's essentially excluded
via CPP. At the same time, it complicates maintenance significantly.
As a result, we decided to have a normal PVP constraint: `< 2.2`.